rename RectifiedCurve to CurvePath and move to animation crate
